magento/adobe-stock-integration#515: Finalized implementation

+++ AdobeStockImage/Model/Storage.php
@@ -97,7 +97,9 @@ class Storage
     public function delete(string $path): void
     {
         try {
-            $result = $this->getMediaDirectory()->delete($path);
+            if ($this->getMediaDirectory()->isFile($path)) {
+                $result = $this->getMediaDirectory()->delete($path);
+            }
         } catch (Exception $exception) {
             $this->log->critical("Failed to delete the image. Exception: \n" . $exception);
         }
